Фигуры вращаются относительно начала. Надо зажать левую клавишу и крутить колесо.
Добавлен слушатель колесика MouseWheelListener, который меняет угол
построения фигуры. Как следствие, фигура вращается относительно одной из
вершин.
Цель -добиться того чтобы фигуры вращались относительно центра. Для
этого для каждого Shape написан свой void rotate(int i), который должен
вращать фигуру относительно центра.

При использовании rotate появились непонятные глюки - фигуры уползают,
но если подвигать мышью, то возвращаются к указателю. Пока
закомментировал в Canvas использование этого метода (134-135 строки)

+++ src/Vector.java
@@ -8,9 +8,9 @@ public class Vector {
 static Vector test = new Vector(3,7);
 
 
-    public Vector(int x, int y) {
-        this.x = x;
-        this.y = y;
+    public Vector(double x, double y) {
+        this.x = (int)x;
+        this.y = (int)y;
     }
 
     public Vector() {
@@ -80,9 +80,9 @@ static Vector test = new Vector(3,7);
     }
 
     //* метод увеличивает вектор в k раз
-    public void times(int k) {
-        x = x * k;
-        y = y * k;
+    public void times(double k) {
+        x = (int)(x * k);
+        y = (int)(y * k);
     }
 
     public String toString() {
